third_party/blink/web_tests/http/tests/websocket/httponly-cookie-expected.txt

Test that WebSocket sends HttpOnly cookies.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS cookie is "ws-httponly=1; ws=1"
PASS successfullyParsed is true

TEST COMPLETE
